requirements.txt
numpy
tensorflow
python-chess
pygame